Pull DeviceConstants into intrinsics header
This makes more sense as intrinsics depend on the device and not
on any of the policies.
